move legacy api into guarded block so it's not included in a server bundle (#1267)
Moves the legacy dependencies api (used by older versions of lasso-marko) into a block that's guarded by !process.env.BUNDLE. APIs inside this block use dynamic requires and are either intended for development or are legacy apis. This is needed so that webpack does not throw warnings or errors when building a bundle for node.